refactor: move isAttribute to type-predicates folder
